last

+++ src/main/java/org/example/Service/cart/CartService.java
@@ -6,7 +6,7 @@ import org.example.model.product.Product;
 import java.util.List;
 
 public interface CartService {
-    void addItem(Product product , Double quantity);
+    void addItem(Product product , Float quantity);
     List<CartItem> getItems();
     boolean isEmpty();
