+++ src/main/java/org/example/Service/shipment/Shippable.java
@@ -2,5 +2,5 @@ package org.example.Service.shipment;
 
 public interface Shippable {
     String getShippableName();
-    Double getWeight();
+    Float getWeight();
 }
